fix: memberId와 providerId를 통합하여 관련 사항 수정 #21
